requirements.txt
streamlit==1.32.2
opencv-python
numpy
Pillow
mediapipe
